add settings danger zone & remove settings project table
